pyproject.toml
[project]
name = "ai-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.2.1",
    "ccxt>=4.5.14",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "ruff>=0.14.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_trader"]

[tool.hatch.build.targets.sdist]
include = [
    "src/ai_trader",
    "examples",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
